+++ runtime/hub/core/authenticators/__init__.py
@@ -25,7 +25,7 @@ Provides various authentication methods for JupyterHub.
 
 from core.authenticators.auto_login import AutoLoginAuthenticator
 from core.authenticators.firstuse import CustomFirstUseAuthenticator
-from core.authenticators.github_oauth import CustomGitHubOAuthenticator
+from core.authenticators.github_app import CustomGitHubOAuthenticator
 from core.authenticators.jwt import RemoteLabAuthenticator
 from core.authenticators.multi import CustomMultiAuthenticator
